Ensure release package has correct dependencies
The dependency list for the release package (.tar.bz2) was very wrong.
It basically contained some Makefiles and the target's binary/library
output.

This updates the ADD_REL_* APIs to track every file that is added to the
release package. That list of files is now used as the release package's
dependency list.