Fix nested directory-level flags on macOS
I'm not entirely sure what the issue here was, but macOS ships with make
version 3.81 (from 2006), and that version had trouble with the way the
directory-level flags were passed to the e.g. $(COMP_CXX) helpers.

Rather than defining the flags as goal-level variables, change those
helpers to be callables and pass the flags as parameters.