Remove static libraries before creating them
The command to create a static library specifies every object file in a
single invocation. If the library already exists, and contains an object
file that was removed from the build, ar will not remove the stale file.
So when creating a static library, remove the old library to ensure no
lingering stale object files are left.