Ensure object files depend on generated sources
In a clean build, this ensures that any header files are generated
before a source file that may include them. The dependency is defined as
order-only to prevent rebuilding the whole world when a generated header
changes. Source files that actually include the generated headers will
explicitly depend on those headers via their dependency (.d) files.